test: 라이더 주문 상태 변경 테스트코드 이동
