+++ src/test/java/com/idukbaduk/itseats/order/controller/RiderOrderControllerTest.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.idukbaduk.itseats.order.dto.OrderDetailsResponse;
 import com.idukbaduk.itseats.order.dto.enums.OrderResponse;
 import com.idukbaduk.itseats.order.service.RiderOrderService;
+import com.idukbaduk.itseats.rider.dto.enums.RiderResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,6 +19,9 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @WebMvcTest(RiderOrderController.class)
 class RiderOrderControllerTest {
@@ -52,5 +56,56 @@ class RiderOrderControllerTest {
                 .andExpect(MockMvcResultMatchers.jsonPath("$.message")
                         .value(OrderResponse.GET_RIDER_ORDER_DETAILS_SUCCESS.getMessage()))
                 .andExpect(MockMvcResultMatchers.jsonPath("$.data.orderId").value(orderId));
+    }
+
+    @Test
+    @DisplayName("배달 수락 성공")
+    @WithMockUser(username = "testuser")
+    void updateDeliveryStatusAccept_success() throws Exception {
+        // given
+        long orderId = 1L;
+
+        // when & then
+        mockMvc.perform(post("/api/rider/" + orderId + "/accept")
+                        .with(csrf()))
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.httpStatus")
+                        .value(RiderResponse.UPDATE_STATUS_ACCEPT_SUCCESS.getHttpStatus().value()))
+                .andExpect(jsonPath("$.message")
+                        .value(RiderResponse.UPDATE_STATUS_ACCEPT_SUCCESS.getMessage()));
+    }
+
+    @Test
+    @DisplayName("배달 완료 성공")
+    @WithMockUser(username = "testuser")
+    void updateDeliveryStatusDone_success() throws Exception {
+        // given
+        long orderId = 1L;
+
+        // when & then
+        mockMvc.perform(post("/api/rider/" + orderId + "/done")
+                        .with(csrf()))
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.httpStatus")
+                        .value(RiderResponse.UPDATE_STATUS_DELIVERED_SUCCESS.getHttpStatus().value()))
+                .andExpect(jsonPath("$.message")
+                        .value(RiderResponse.UPDATE_STATUS_DELIVERED_SUCCESS.getMessage()));
+    }
+
+    @Test
+    @DisplayName("픽업 완료 성공")
+    @WithMockUser(username = "testuser")
+    void updateDeliveryStatusPickup_success() throws Exception {
+        // given
+        long orderId = 1L;
+
+        // when & then
+        mockMvc.perform(post("/api/rider/" + orderId + "/pickup")
+                        .with(csrf()))
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.httpStatus")
+                        .value(RiderResponse.UPDATE_STATUS_PICKUP_SUCCESS.getHttpStatus().value()))
+                .andExpect(jsonPath("$.message")
+                        .value(RiderResponse.UPDATE_STATUS_PICKUP_SUCCESS.getMessage()));
     }
 }
