+++ src/test/java/com/idukbaduk/itseats/order/service/RiderOrderServiceTest.java
@@ -184,4 +184,140 @@ class RiderOrderServiceTest {
                 .isInstanceOf(OrderException.class)
                 .hasMessageContaining(OrderErrorCode.ORDER_NOT_FOUND.getMessage());
     }
+
+    @Test
+    @DisplayName("주문 상태를 배차 완료 상태로 변경에 성공")
+    void acceptDelivery_success() {
+        // given
+        Order order = Order.builder()
+                .orderId(1L)
+                .orderStatus(OrderStatus.COOKED)
+                .build();
+
+        when(memberRepository.findByUsername(username)).thenReturn(Optional.of(member));
+        when(riderRepository.findByMember(member)).thenReturn(Optional.of(rider));
+        when(orderRepository.findById(1L)).thenReturn(Optional.of(order));
+
+        // when
+        riderOrderService.acceptDelivery(username, 1L);
+
+        // then
+        assertThat(order.getOrderId()).isEqualTo(1L);
+        assertThat(order.getRider()).isEqualTo(rider);
+        assertThat(order.getOrderStatus()).isEqualTo(OrderStatus.RIDER_READY);
+    }
+
+    @Test
+    @DisplayName("주문이 직전 단계가 아닌 경우 주문 상태를 변경하면 예외 발생")
+    void acceptDelivery_acceptStatusFail() {
+        // given
+        Order order = Order.builder()
+                .orderId(1L)
+                .orderStatus(OrderStatus.COOKING)
+                .build();
+
+        when(memberRepository.findByUsername(username)).thenReturn(Optional.of(member));
+        when(riderRepository.findByMember(member)).thenReturn(Optional.of(rider));
+        when(orderRepository.findById(1L)).thenReturn(Optional.of(order));
+
+        // when
+        assertThatThrownBy(() -> riderOrderService.acceptDelivery(username, 1L))
+                .isInstanceOf(OrderException.class)
+                .hasMessageContaining(OrderErrorCode.ORDER_STATUS_UPDATE_FAIL.getMessage());
+    }
+
+    @Test
+    @DisplayName("라이더가 이미 배차된 주문의 주문 상태를 변경하면 예외 발생")
+    void acceptDelivery_alreadyAssignedRider() {
+        // given
+        Order order = Order.builder()
+                .orderId(1L)
+                .orderStatus(OrderStatus.COOKED)
+                .rider(rider)
+                .build();
+
+        when(memberRepository.findByUsername(username)).thenReturn(Optional.of(member));
+        when(riderRepository.findByMember(member)).thenReturn(Optional.of(rider));
+        when(orderRepository.findById(1L)).thenReturn(Optional.of(order));
+
+        // when
+        assertThatThrownBy(() -> riderOrderService.acceptDelivery(username, 1L))
+                .isInstanceOf(OrderException.class)
+                .hasMessageContaining(OrderErrorCode.ORDER_ALREADY_ASSIGNED.getMessage());
+    }
+
+    @Test
+    @DisplayName("주문 상태를 배달 완료 상태로 변경에 성공")
+    void updateOrderStatusAfterAccept_delivered_success() {
+        // given
+        Order order = Order.builder()
+                .orderId(1L)
+                .orderStatus(OrderStatus.DELIVERING)
+                .rider(rider)
+                .build();
+
+        when(memberRepository.findByUsername(username)).thenReturn(Optional.of(member));
+        when(riderRepository.findByMember(member)).thenReturn(Optional.of(rider));
+        when(orderRepository.findByRiderAndOrderId(rider, 1L)).thenReturn(Optional.of(order));
+
+        // when
+        riderOrderService.updateOrderStatusAfterAccept(username, 1L, OrderStatus.DELIVERED);
+
+        // then
+        assertThat(order.getOrderStatus()).isEqualTo(OrderStatus.DELIVERED);
+    }
+
+    @Test
+    @DisplayName("주문이 직전 단계가 아닌 경우 주문 상태를 변경하면 예외 발생")
+    void updateOrderStatusAfterAccept_deliveredFail() {
+        // given
+        Order order = Order.builder()
+                .orderId(1L)
+                .orderStatus(OrderStatus.COOKING)
+                .build();
+
+        when(memberRepository.findByUsername(username)).thenReturn(Optional.of(member));
+        when(riderRepository.findByMember(member)).thenReturn(Optional.of(rider));
+        when(orderRepository.findByRiderAndOrderId(rider, 1L)).thenReturn(Optional.of(order));
+
+        // when
+        assertThatThrownBy(() -> riderOrderService.updateOrderStatusAfterAccept(username, 1L, OrderStatus.DELIVERED))
+                .isInstanceOf(OrderException.class)
+                .hasMessageContaining(OrderErrorCode.ORDER_STATUS_UPDATE_FAIL.getMessage());
+    }
+
+    @Test
+    @DisplayName("배차되지 않은 주문의 상태를 변경하는 경우 예외 발생")
+    void updateOrderStatusAfterAccept_getOrderFail() {
+        // given
+        when(memberRepository.findByUsername(username)).thenReturn(Optional.of(member));
+        when(riderRepository.findByMember(member)).thenReturn(Optional.of(rider));
+        when(orderRepository.findByRiderAndOrderId(rider, 1L)).thenReturn(Optional.empty());
+
+        // when & then
+        assertThatThrownBy(() -> riderOrderService.updateOrderStatusAfterAccept(username, 1L, OrderStatus.DELIVERED))
+                .isInstanceOf(OrderException.class)
+                .hasMessageContaining(OrderErrorCode.ORDER_NOT_FOUND.getMessage());
+    }
+
+    @Test
+    @DisplayName("주문 상태를 배달 중 상태로 변경에 성공")
+    void updateOrderStatusAfterAccept_delivering_success() {
+        // given
+        Order order = Order.builder()
+                .orderId(1L)
+                .orderStatus(OrderStatus.RIDER_READY)
+                .rider(rider)
+                .build();
+
+        when(memberRepository.findByUsername(username)).thenReturn(Optional.of(member));
+        when(riderRepository.findByMember(member)).thenReturn(Optional.of(rider));
+        when(orderRepository.findByRiderAndOrderId(rider, 1L)).thenReturn(Optional.of(order));
+
+        // when
+        riderOrderService.updateOrderStatusAfterAccept(username, 1L, OrderStatus.DELIVERING);
+
+        // then
+        assertThat(order.getOrderStatus()).isEqualTo(OrderStatus.DELIVERING);
+    }
 }
