+++ src/test/java/com/idukbaduk/itseats/rider/controller/RiderControllerTest.java
@@ -74,55 +74,4 @@ class RiderControllerTest {
                 .andExpect(status().isBadRequest())
                 .andExpect(jsonPath("$.message").value("출/퇴근 여부는 필수값입니다."));
     }
-
-    @Test
-    @DisplayName("배달 수락 성공")
-    @WithMockUser(username = "testuser")
-    void updateDeliveryStatusAccept_success() throws Exception {
-        // given
-        long orderId = 1L;
-
-        // when & then
-        mockMvc.perform(post("/api/rider/" + orderId + "/accept")
-                        .with(csrf()))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.httpStatus")
-                        .value(RiderResponse.UPDATE_STATUS_ACCEPT_SUCCESS.getHttpStatus().value()))
-                .andExpect(jsonPath("$.message")
-                        .value(RiderResponse.UPDATE_STATUS_ACCEPT_SUCCESS.getMessage()));
-    }
-
-    @Test
-    @DisplayName("배달 완료 성공")
-    @WithMockUser(username = "testuser")
-    void updateDeliveryStatusDone_success() throws Exception {
-        // given
-        long orderId = 1L;
-
-        // when & then
-        mockMvc.perform(post("/api/rider/" + orderId + "/done")
-                        .with(csrf()))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.httpStatus")
-                        .value(RiderResponse.UPDATE_STATUS_DELIVERED_SUCCESS.getHttpStatus().value()))
-                .andExpect(jsonPath("$.message")
-                        .value(RiderResponse.UPDATE_STATUS_DELIVERED_SUCCESS.getMessage()));
-    }
-
-    @Test
-    @DisplayName("픽업 완료 성공")
-    @WithMockUser(username = "testuser")
-    void updateDeliveryStatusPickup_success() throws Exception {
-        // given
-        long orderId = 1L;
-
-        // when & then
-        mockMvc.perform(post("/api/rider/" + orderId + "/pickup")
-                        .with(csrf()))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.httpStatus")
-                        .value(RiderResponse.UPDATE_STATUS_PICKUP_SUCCESS.getHttpStatus().value()))
-                .andExpect(jsonPath("$.message")
-                        .value(RiderResponse.UPDATE_STATUS_PICKUP_SUCCESS.getMessage()));
-    }
 }
