+++ src/test/java/com/idukbaduk/itseats/rider/service/RiderServiceTest.java
@@ -2,10 +2,6 @@ package com.idukbaduk.itseats.rider.service;
 
 import com.idukbaduk.itseats.member.entity.Member;
 import com.idukbaduk.itseats.member.repository.MemberRepository;
-import com.idukbaduk.itseats.order.entity.Order;
-import com.idukbaduk.itseats.order.entity.enums.OrderStatus;
-import com.idukbaduk.itseats.order.error.OrderException;
-import com.idukbaduk.itseats.order.error.enums.OrderErrorCode;
 import com.idukbaduk.itseats.order.repository.OrderRepository;
 import com.idukbaduk.itseats.rider.dto.ModifyWorkingRequest;
 import com.idukbaduk.itseats.rider.dto.WorkingInfoResponse;
@@ -34,8 +30,6 @@ class RiderServiceTest {
     private RiderRepository riderRepository;
     @Mock
     private MemberRepository memberRepository;
-    @Mock
-    private OrderRepository orderRepository;
 
     @InjectMocks
     private RiderService riderService;
@@ -104,141 +98,5 @@ class RiderServiceTest {
         assertThatThrownBy(() -> riderService.modifyWorking(username, ModifyWorkingRequest.builder().build()))
                 .isInstanceOf(RiderException.class)
                 .hasMessageContaining(RiderErrorCode.RIDER_NOT_FOUND.getMessage());
-    }
-
-    @Test
-    @DisplayName("주문 상태를 배차 완료 상태로 변경에 성공")
-    void acceptDelivery_success() {
-        // given
-        Order order = Order.builder()
-                .orderId(1L)
-                .orderStatus(OrderStatus.COOKED)
-                .build();
-
-        when(memberRepository.findByUsername(username)).thenReturn(Optional.of(member));
-        when(riderRepository.findByMember(member)).thenReturn(Optional.of(rider));
-        when(orderRepository.findById(1L)).thenReturn(Optional.of(order));
-
-        // when
-        riderService.acceptDelivery(username, 1L);
-
-        // then
-        assertThat(order.getOrderId()).isEqualTo(1L);
-        assertThat(order.getRider()).isEqualTo(rider);
-        assertThat(order.getOrderStatus()).isEqualTo(OrderStatus.RIDER_READY);
-    }
-
-    @Test
-    @DisplayName("주문이 직전 단계가 아닌 경우 주문 상태를 변경하면 예외 발생")
-    void acceptDelivery_acceptStatusFail() {
-        // given
-        Order order = Order.builder()
-                .orderId(1L)
-                .orderStatus(OrderStatus.COOKING)
-                .build();
-
-        when(memberRepository.findByUsername(username)).thenReturn(Optional.of(member));
-        when(riderRepository.findByMember(member)).thenReturn(Optional.of(rider));
-        when(orderRepository.findById(1L)).thenReturn(Optional.of(order));
-
-        // when
-        assertThatThrownBy(() -> riderService.acceptDelivery(username, 1L))
-                .isInstanceOf(OrderException.class)
-                .hasMessageContaining(OrderErrorCode.ORDER_STATUS_UPDATE_FAIL.getMessage());
-    }
-
-    @Test
-    @DisplayName("라이더가 이미 배차된 주문의 주문 상태를 변경하면 예외 발생")
-    void acceptDelivery_alreadyAssignedRider() {
-        // given
-        Order order = Order.builder()
-                .orderId(1L)
-                .orderStatus(OrderStatus.COOKED)
-                .rider(rider)
-                .build();
-
-        when(memberRepository.findByUsername(username)).thenReturn(Optional.of(member));
-        when(riderRepository.findByMember(member)).thenReturn(Optional.of(rider));
-        when(orderRepository.findById(1L)).thenReturn(Optional.of(order));
-
-        // when
-        assertThatThrownBy(() -> riderService.acceptDelivery(username, 1L))
-                .isInstanceOf(OrderException.class)
-                .hasMessageContaining(OrderErrorCode.ORDER_ALREADY_ASSIGNED.getMessage());
-    }
-
-    @Test
-    @DisplayName("주문 상태를 배달 완료 상태로 변경에 성공")
-    void updateOrderStatusAfterAccept_delivered_success() {
-        // given
-        Order order = Order.builder()
-                .orderId(1L)
-                .orderStatus(OrderStatus.DELIVERING)
-                .rider(rider)
-                .build();
-
-        when(memberRepository.findByUsername(username)).thenReturn(Optional.of(member));
-        when(riderRepository.findByMember(member)).thenReturn(Optional.of(rider));
-        when(orderRepository.findByRiderAndOrderId(rider, 1L)).thenReturn(Optional.of(order));
-
-        // when
-        riderService.updateOrderStatusAfterAccept(username, 1L, OrderStatus.DELIVERED);
-
-        // then
-        assertThat(order.getOrderStatus()).isEqualTo(OrderStatus.DELIVERED);
-    }
-
-    @Test
-    @DisplayName("주문이 직전 단계가 아닌 경우 주문 상태를 변경하면 예외 발생")
-    void updateOrderStatusAfterAccept_deliveredFail() {
-        // given
-        Order order = Order.builder()
-                .orderId(1L)
-                .orderStatus(OrderStatus.COOKING)
-                .build();
-
-        when(memberRepository.findByUsername(username)).thenReturn(Optional.of(member));
-        when(riderRepository.findByMember(member)).thenReturn(Optional.of(rider));
-        when(orderRepository.findByRiderAndOrderId(rider, 1L)).thenReturn(Optional.of(order));
-
-        // when
-        assertThatThrownBy(() -> riderService.updateOrderStatusAfterAccept(username, 1L, OrderStatus.DELIVERED))
-                .isInstanceOf(OrderException.class)
-                .hasMessageContaining(OrderErrorCode.ORDER_STATUS_UPDATE_FAIL.getMessage());
-    }
-
-    @Test
-    @DisplayName("배차되지 않은 주문의 상태를 변경하는 경우 예외 발생")
-    void updateOrderStatusAfterAccept_getOrderFail() {
-        // given
-        when(memberRepository.findByUsername(username)).thenReturn(Optional.of(member));
-        when(riderRepository.findByMember(member)).thenReturn(Optional.of(rider));
-        when(orderRepository.findByRiderAndOrderId(rider, 1L)).thenReturn(Optional.empty());
-
-        // when & then
-        assertThatThrownBy(() -> riderService.updateOrderStatusAfterAccept(username, 1L, OrderStatus.DELIVERED))
-                .isInstanceOf(OrderException.class)
-                .hasMessageContaining(OrderErrorCode.ORDER_NOT_FOUND.getMessage());
-    }
-
-    @Test
-    @DisplayName("주문 상태를 배달 중 상태로 변경에 성공")
-    void updateOrderStatusAfterAccept_delivering_success() {
-        // given
-        Order order = Order.builder()
-                .orderId(1L)
-                .orderStatus(OrderStatus.RIDER_READY)
-                .rider(rider)
-                .build();
-
-        when(memberRepository.findByUsername(username)).thenReturn(Optional.of(member));
-        when(riderRepository.findByMember(member)).thenReturn(Optional.of(rider));
-        when(orderRepository.findByRiderAndOrderId(rider, 1L)).thenReturn(Optional.of(order));
-
-        // when
-        riderService.updateOrderStatusAfterAccept(username, 1L, OrderStatus.DELIVERING);
-
-        // then
-        assertThat(order.getOrderStatus()).isEqualTo(OrderStatus.DELIVERING);
     }
 }
